Conflicts:
	common/glfw_view.cpp
	include/mbgl/map/map.hpp
	include/mbgl/map/transform.hpp
	ios/mapbox-gl-cocoa
	src/map/map.cpp
	src/mbgl/map/transform.cpp
	styles
	test/headless.cpp

+++ src/libtess2/bucketalloc.c
@@ -35,7 +35,7 @@ extern "C" {
 
 #include <stdio.h>
 #include <stdlib.h>
-#include <libtess2/tesselator.h>
+#include "tesselator.h"
 
 //#define CHECK_BOUNDS
 
@@ -59,7 +59,7 @@ struct BucketAlloc
 
 static int CreateBucket( struct BucketAlloc* ba )
 {
-	size_t size;
+	unsigned int size;
 	Bucket* bucket;
 	void* freelist;
 	unsigned char* head;
